+++ cpp/subprojects/boosting/src/mlrl/boosting/rule_evaluation/regularization_no.cpp
@@ -2,8 +2,8 @@
 
 namespace boosting {
 
-    float64 NoRegularizationConfig::getWeight() const {
-        return 0;
+    float32 NoRegularizationConfig::getWeight() const {
+        return 0.0f;
     }
 
 }
